+++ ecs/pool_test.go
@@ -133,3 +133,22 @@ func TestBitPool(t *testing.T) {
 		assert.Equal(t, i, int(p.Get()))
 	}
 }
+
+func TestIntPool(t *testing.T) {
+	p := newIntPool[int](16)
+
+	for n := 0; n < 3; n++ {
+		for i := 0; i < 32; i++ {
+			assert.Equal(t, i, p.Get())
+		}
+
+		assert.Equal(t, 32, len(p.pool))
+
+		p.Recycle(3)
+		p.Recycle(4)
+		assert.Equal(t, 4, p.Get())
+		assert.Equal(t, 3, p.Get())
+
+		p.Reset()
+	}
+}